Small refactoring of handlers
