+++ src/com/maddyhome/idea/vim/handler/ChangeEditorActionHandler.kt
@@ -29,17 +29,35 @@ import com.maddyhome.idea.vim.command.Command
 import com.maddyhome.idea.vim.helper.vimChangeActionSwitchMode
 import com.maddyhome.idea.vim.helper.vimLastColumn
 
-sealed class ChangeEditorActionHandler : VimActionHandler.SingleExecution() {
+/**
+ * Base handler for commands that performs change actions.
+ * This handler stores the commands and they can be repeated later with dot command.
+ *
+ * Use subclasses of this handler:
+ *   - [ChangeEditorActionHandler.SingleExecution]
+ *   - [ChangeEditorActionHandler.ForEachCaret]
+ */
+sealed class ChangeEditorActionHandler : EditorActionHandlerBase(false) {
 
+  /**
+   * This handler executes an action for each caret. That means that if you have 5 carets, [execute] will be
+   *   called 5 times.
+   * @see [ChangeEditorActionHandler.SingleExecution] for only one execution.
+   */
   abstract class ForEachCaret : ChangeEditorActionHandler() {
     abstract fun execute(editor: Editor, caret: Caret, context: DataContext, count: Int, rawCount: Int, argument: Argument?): Boolean
   }
 
+  /**
+   * This handler executes an action only once for all carets. That means that if you have 5 carets,
+   *   [execute] will be called 1 time.
+   * @see [ChangeEditorActionHandler.ForEachCaret] for per-caret execution
+   */
   abstract class SingleExecution : ChangeEditorActionHandler() {
     abstract fun execute(editor: Editor, context: DataContext, count: Int, rawCount: Int, argument: Argument?): Boolean
   }
 
-  final override fun execute(editor: Editor, context: DataContext, cmd: Command): Boolean {
+  final override fun baseExecute(editor: Editor, caret: Caret, context: DataContext, cmd: Command): Boolean {
     // Here we have to save the last changed command. This should be done separately for each
     // call of the task, not for each caret. Currently there is no way to schedule any action
     // to be worked after each task. So here we override the deprecated execute function which
@@ -50,9 +68,9 @@ sealed class ChangeEditorActionHandler : VimActionHandler.SingleExecution() {
     val worked = Ref.create(true)
     when (this) {
       is ForEachCaret -> {
-        editor.caretModel.runForEachCaret({ caret ->
-          if (!caret.isValid) return@runForEachCaret
-          if (!execute(editor, caret, context, cmd.count, cmd.rawCount, cmd.argument)) {
+        editor.caretModel.runForEachCaret({ current ->
+          if (!current.isValid) return@runForEachCaret
+          if (!execute(editor, current, context, cmd.count, cmd.rawCount, cmd.argument)) {
             worked.set(false)
           }
         }, true)
