+++ src/com/maddyhome/idea/vim/handler/VisualOperatorActionHandler.kt
@@ -52,14 +52,17 @@ import com.maddyhome.idea.vim.helper.vimSelectionStart
  * @author Alex Plate
  *
  * Base class for visual operation handlers.
- * @see [VisualOperatorActionHandler.SingleExecution] and [VisualOperatorActionHandler.ForEachCaret]
+ *
+ * Use subclasses of this handler:
+ *  - [VisualOperatorActionHandler.SingleExecution]
+ *  - [VisualOperatorActionHandler.ForEachCaret]
  */
-sealed class VisualOperatorActionHandler : VimActionHandler.SingleExecution() {
+sealed class VisualOperatorActionHandler : EditorActionHandlerBase(false) {
   /**
    * Base class for visual operation handlers.
    * This handler executes an action for each caret. That means that if you have 5 carets,
    *   [executeAction] will be called 5 times.
-   * @see [VisualOperatorActionHandler.SingleExecution] for only one execution
+   * @see [VisualOperatorActionHandler.SingleExecution] for only one execution.
    */
   abstract class ForEachCaret : VisualOperatorActionHandler() {
 
@@ -90,7 +93,7 @@ sealed class VisualOperatorActionHandler : VimActionHandler.SingleExecution() {
   /**
    * Base class for visual operation handlers.
    * This handler executes an action only once for all carets. That means that if you have 5 carets,
-   *   [executeAction] will be called 1 time.
+   *   [executeForAllCarets] will be called 1 time.
    * @see [VisualOperatorActionHandler.ForEachCaret] for per-caret execution
    */
   abstract class SingleExecution : VisualOperatorActionHandler() {
@@ -104,7 +107,7 @@ sealed class VisualOperatorActionHandler : VimActionHandler.SingleExecution() {
     abstract fun executeForAllCarets(editor: Editor, context: DataContext, cmd: Command, caretsAndSelections: Map<Caret, VimSelection>): Boolean
   }
 
-  final override fun execute(editor: Editor, context: DataContext, cmd: Command): Boolean {
+  final override fun baseExecute(editor: Editor, caret: Caret, context: DataContext, cmd: Command): Boolean {
     logger.info("Execute visual command $cmd")
 
     editor.vimChangeActionSwitchMode = null
@@ -131,9 +134,9 @@ sealed class VisualOperatorActionHandler : VimActionHandler.SingleExecution() {
         when {
           selections.keys.isEmpty() -> return false
           selections.keys.size == 1 -> res.set(executeAction(editor, selections.keys.first(), context, cmd, selections.values.first()))
-          else -> editor.caretModel.runForEachCaret({ caret ->
-            val range = selections.getValue(caret)
-            val loopRes = executeAction(editor, caret, context, cmd, range)
+          else -> editor.caretModel.runForEachCaret({ currentCaret ->
+            val range = selections.getValue(currentCaret)
+            val loopRes = executeAction(editor, currentCaret, context, cmd, range)
             res.set(loopRes and res.get())
           }, true)
         }
